+++ src/Utility/Checkups/Directories.php
@@ -11,15 +11,17 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       2.22.8
+ * @since       2.29.0
  */
 
 namespace MeCms\Utility\Checkups;
 
+use Cake\Core\Configure;
+
 /**
- * Checkup for webroot directories
+ * Checkup for directories
  */
-class Webroot
+class Directories
 {
     /**
      * Checks if each path is writeable
@@ -27,15 +29,12 @@ class Webroot
      */
     public static function isWriteable(): array
     {
-        foreach ([
-            BANNERS,
-            PHOTOS,
-            USER_PICTURES,
-            UPLOADED,
-        ] as $path) {
+        foreach (Configure::read('WRITABLE_DIRS') as $path) {
             $result[$path] = is_writable_resursive($path);
         }
 
-        return $result ?? [];
+        ksort($result);
+
+        return $result;
     }
 }